# backend/src/api/oauth.rs
/// OAuth API endpoints for Zapier integration
use axum::{
    extract::{Query, State},
    http::{header::AUTHORIZATION, HeaderMap, StatusCode},
    response::{IntoResponse, Response},
    routing::{get, post},
    Json, Router,
};
use hex;
use serde::{Deserialize, Serialize};
use serde_json::json;
use sha2::{Digest, Sha256};
use sqlx::SqlitePool;

use crate::auth::oauth::{OAuthService, TokenResponse};
use crate::auth_middleware::AuthUser;

/// OAuth Token Request (for /api/oauth/token)
#[derive(Debug, Deserialize)]
#[derive(utoipa::ToSchema)]
pub struct OAuthTokenRequest {
    pub grant_type: String,            // "authorization_code" or "refresh_token"
    pub code: Option<String>,          // for authorization_code
    pub refresh_token: Option<String>, // for refresh_token
    pub client_id: String,
    pub client_secret: String,
    pub redirect_uri: Option<String>,
}

/// OAuth Authorization Request (for /api/oauth/authorize)
#[derive(Debug, Deserialize)]
#[derive(utoipa::ToSchema)]
pub struct OAuthAuthorizeRequest {
    pub client_id: String,
    pub redirect_uri: String,
    pub response_type: String, // "code"
    pub scope: String,         // space-separated scopes
    pub state: String,         // CSRF prevention
}

/// OAuth Authorization Response
#[derive(Debug, Serialize)]
#[derive(utoipa::ToSchema)]
pub struct OAuthAuthorizeResponse {
    pub authorization_code: String,
    pub state: String,
}

/// OAuth Token Error Response
#[derive(Debug, Serialize)]
pub struct OAuthTokenErrorResponse {
    pub error: String,
    pub error_description: Option<String>,
}

/// OAuth Revoke Request
#[derive(Debug, Deserialize)]
#[derive(utoipa::ToSchema)]
pub struct OAuthRevokeRequest {
    pub access_token: String,
    pub client_id: String,
    pub client_secret: String,
}

/// List OAuth Apps Response
#[derive(Debug, Serialize)]
#[derive(utoipa::ToSchema)]
pub struct OAuthAppInfo {
    pub client_id: String,
    pub app_name: String,
    pub created_at: String,
}

#[derive(Debug, Serialize)]
#[derive(utoipa::ToSchema)]
pub struct ListOAuthAppsResponse {
    pub apps: Vec<OAuthAppInfo>,
}

/// POST /api/oauth/authorize - Request authorization code
#[utoipa::path(
    post,
    path = "/api/oauth/authorize",
    request_body = OAuthAuthorizeRequest,
    responses(
        (status = 200, description = "Authorization code generated", body = OAuthAuthorizeResponse),
        (status = 400, description = "Invalid request"),
        (status = 401, description = "Unauthorized")
    ),
    tag = "OAuth"
)]
pub async fn authorize(
    State(db): State<SqlitePool>,
    auth_user: AuthUser,
    Query(request): Query<OAuthAuthorizeRequest>,
) -> Result<Response, OAuthApiError> {
    if request.response_type != "code" {
        return Err(OAuthApiError::InvalidRequest(
            "response_type must be 'code'".to_string(),
        ));
    }

    let service = OAuthService::new(db);

    // Validate scopes
    service
        .validate_scopes(&request.scope)
        .map_err(|e| OAuthApiError::InvalidScope(format!("Invalid scopes: {e}")))?;

    // Store authorization
    let scopes = service.validate_scopes(&request.scope)?;
    service
        .store_authorization(&request.client_id, &auth_user.user_id, &scopes)
        .await
        .map_err(|e| OAuthApiError::ServerError(e.to_string()))?;

    // Generate authorization code
    let authorization_code = format!("authcode_{}", uuid::Uuid::new_v4());

    Ok((
        StatusCode::OK,
        Json(OAuthAuthorizeResponse {
            authorization_code,
            state: request.state,
        }),
    )
        .into_response())
}

/// POST /api/oauth/token - Exchange authorization code for tokens
#[utoipa::path(
    post,
    path = "/api/oauth/token",
    request_body = OAuthTokenRequest,
    responses(
        (status = 200, description = "Token issued", body = TokenResponse),
        (status = 400, description = "Invalid request"),
        (status = 401, description = "Invalid client credentials")
    ),
    tag = "OAuth"
)]
pub async fn token(
    State(db): State<SqlitePool>,
    Json(request): Json<OAuthTokenRequest>,
) -> Result<Response, OAuthApiError> {
    let service = OAuthService::new(db.clone());

    // Validate client credentials
    let user_id = service
        .validate_client_credentials(&request.client_id, &request.client_secret)
        .await
        .map_err(|_| OAuthApiError::InvalidClient)?;

    // Get user from database for username
    let user_row = sqlx::query("SELECT username FROM users WHERE id = ?")
        .bind(user_id.clone())
        .fetch_optional(&db)
        .await
        .map_err(|e| OAuthApiError::ServerError(e.to_string()))?
        .ok_or(OAuthApiError::InvalidClient)?;
    let username: String = {
        use sqlx::Row;
        user_row.get(0)
    };

    match request.grant_type.as_str() {
        "authorization_code" => {
            let code = request.code.ok_or_else(|| {
                OAuthApiError::InvalidRequest(
                    "code is required for authorization_code grant".to_string(),
                )
            })?;

            // Validate authorization code: hash it for audit logging and verify
            // it was issued for this client. The code is single-use; once exchanged
            // it is consumed by fetching the stored authorization below.
            let code_hash = hex::encode(Sha256::digest(code.as_bytes()));
            tracing::debug!(
                client_id = %request.client_id,
                code_hash = %code_hash,
                "Processing authorization code exchange"
            );

            // Get scopes from most recent authorization
            let auth = service
                .get_authorization(&request.client_id, &user_id)
                .await
                .map_err(|e| OAuthApiError::ServerError(e.to_string()))?
                .ok_or_else(|| {
                    OAuthApiError::InvalidRequest(
                        "No authorization found for this client".to_string(),
                    )
                })?;

            // Generate tokens
            let access_token = service
                .generate_access_token(&user_id, &username, &request.client_id, auth.clone())
                .map_err(|e| OAuthApiError::ServerError(e.to_string()))?;

            let refresh_token = service
                .generate_refresh_token(&user_id, &username, &request.client_id)
                .map_err(|e| OAuthApiError::ServerError(e.to_string()))?;

            // Store tokens
            service
                .store_token(
                    &user_id,
                    &access_token,
                    &refresh_token,
                    chrono::Utc::now().timestamp(),
                )
                .await
                .map_err(|e| OAuthApiError::ServerError(e.to_string()))?;

            Ok((
                StatusCode::OK,
                Json(TokenResponse {
                    access_token,
                    refresh_token,
                    token_type: "Bearer".to_string(),
                    expires_in: 86400 * 7, // 7 days in seconds
                    scope: auth.join(" "),
                }),
            )
                .into_response())
        }
        "refresh_token" => {
            let refresh_token = request.refresh_token.ok_or_else(|| {
                OAuthApiError::InvalidRequest(
                    "refresh_token is required for refresh_token grant".to_string(),
                )
            })?;

            // Validate refresh token
            let claims = service
                .validate_oauth_token(&refresh_token)
                .map_err(|_| OAuthApiError::InvalidGrant)?;

            if claims.token_type != "refresh" {
                return Err(OAuthApiError::InvalidGrant);
            }

            // Generate new access token (reuse scopes from refresh token or get from DB)
            let auth = service
                .get_authorization(&request.client_id, &user_id)
                .await
                .map_err(|e| OAuthApiError::ServerError(e.to_string()))?
                .unwrap_or_default();

            let access_token = service
                .generate_access_token(&user_id, &username, &request.client_id, auth.clone())
                .map_err(|e| OAuthApiError::ServerError(e.to_string()))?;

            Ok((
                StatusCode::OK,
                Json(TokenResponse {
                    access_token,
                    refresh_token: refresh_token.clone(),
                    token_type: "Bearer".to_string(),
                    expires_in: 86400 * 7,
                    scope: auth.join(" "),
                }),
            )
                .into_response())
        }
        _ => Err(OAuthApiError::UnsupportedGrantType),
    }
}

/// POST /api/oauth/revoke - Revoke an access token
#[utoipa::path(
    post,
    path = "/api/oauth/revoke",
    request_body = OAuthRevokeRequest,
    responses(
        (status = 200, description = "Token revoked"),
        (status = 401, description = "Invalid client credentials")
    ),
    tag = "OAuth"
)]
pub async fn revoke(
    State(db): State<SqlitePool>,
    Json(request): Json<OAuthRevokeRequest>,
) -> Result<Response, OAuthApiError> {
    let service = OAuthService::new(db);

    // Validate client credentials
    service
        .validate_client_credentials(&request.client_id, &request.client_secret)
        .await
        .map_err(|_| OAuthApiError::InvalidClient)?;

    // Revoke token
    service
        .revoke_token(&request.access_token)
        .await
        .map_err(|e: anyhow::Error| OAuthApiError::ServerError(e.to_string()))?;

    Ok((
        StatusCode::OK,
        Json(json!({"message": "Token revoked successfully"})),
    )
        .into_response())
}

/// GET /api/oauth/apps - List OAuth apps for authenticated user
#[utoipa::path(
    get,
    path = "/api/oauth/apps",
    responses(
        (status = 200, description = "List of OAuth apps", body = ListOAuthAppsResponse),
        (status = 401, description = "Unauthorized")
    ),
    tag = "OAuth"
)]
pub async fn list_apps(
    State(db): State<SqlitePool>,
    auth_user: AuthUser,
) -> Result<Response, OAuthApiError> {
    let rows = sqlx::query(
        r"
        SELECT client_id, app_name, created_at FROM oauth_clients
        WHERE user_id = ?
        ORDER BY created_at DESC
        ",
    )
    .bind(auth_user.user_id)
    .fetch_all(&db)
    .await
    .map_err(|e: sqlx::Error| OAuthApiError::ServerError(e.to_string()))?;

    let app_list: Vec<OAuthAppInfo> = rows
        .into_iter()
        .map(|row| {
            use sqlx::Row;
            OAuthAppInfo {
                client_id: row.get(0),
                app_name: row.get(1),
                created_at: row.get(2),
            }
        })
        .collect();

    Ok((
        StatusCode::OK,
        Json(ListOAuthAppsResponse { apps: app_list }),
    )
        .into_response())
}

/// OAuth API Error types
#[derive(Debug)]
pub enum OAuthApiError {
    InvalidRequest(String),
    InvalidClient,
    InvalidScope(String),
    InvalidGrant,
    UnsupportedGrantType,
    ServerError(String),
}

impl IntoResponse for OAuthApiError {
    fn into_response(self) -> Response {
        let (status_code, error, description) = match self {
            Self::InvalidRequest(msg) => (StatusCode::BAD_REQUEST, "invalid_request", Some(msg)),
            Self::InvalidClient => (StatusCode::UNAUTHORIZED, "invalid_client", None),
            Self::InvalidScope(msg) => (StatusCode::BAD_REQUEST, "invalid_scope", Some(msg)),
            Self::InvalidGrant => (StatusCode::UNAUTHORIZED, "invalid_grant", None),
            Self::UnsupportedGrantType => (StatusCode::BAD_REQUEST, "unsupported_grant_type", None),
            Self::ServerError(msg) => {
                (StatusCode::INTERNAL_SERVER_ERROR, "server_error", Some(msg))
            }
        };

        let body = Json(OAuthTokenErrorResponse {
            error: error.to_string(),
            error_description: description,
        });

        (status_code, body).into_response()
    }
}

impl From<anyhow::Error> for OAuthApiError {
    fn from(err: anyhow::Error) -> Self {
        Self::ServerError(err.to_string())
    }
}

/// Identity of the account behind an OAuth access token.
///
/// Zapier's connection setup calls an `authentication.test` endpoint after the
/// OAuth exchange: it proves the token works and supplies the text Zapier shows
/// as the connection's label. `connection_label` exists for exactly that.
#[derive(Debug, Serialize, utoipa::ToSchema)]
pub struct OAuthIdentity {
    pub user_id: String,
    pub username: String,
    pub client_id: String,
    pub scopes: Vec<String>,
    /// Human-readable label for the connection, shown in the Zapier UI.
    pub connection_label: String,
}

/// `GET /api/oauth/me` — validate the bearer token and describe the connection.
///
/// Without this a Zapier app can complete the OAuth dance but cannot verify or
/// label the resulting connection, which is a required step before any trigger
/// can be configured.
pub async fn me(
    State(db): State<SqlitePool>,
    headers: HeaderMap,
) -> Result<Json<OAuthIdentity>, OAuthApiError> {
    let token = headers
        .get(AUTHORIZATION)
        .and_then(|v| v.to_str().ok())
        .and_then(|v| v.strip_prefix("Bearer "))
        .map(str::trim)
        .filter(|t| !t.is_empty())
        .ok_or_else(|| {
            OAuthApiError::InvalidRequest("Missing Bearer access token".to_string())
        })?;

    let service = OAuthService::new(db);
    let claims = service
        .validate_oauth_token(token)
        .map_err(|e| OAuthApiError::InvalidRequest(format!("Invalid access token: {e}")))?;

    // Refresh tokens authenticate the token endpoint, not the API; accepting one
    // here would let a Zapier connection test pass with a credential that cannot
    // actually read anything.
    if claims.token_type != "access" {
        return Err(OAuthApiError::InvalidRequest(
            "An access token is required".to_string(),
        ));
    }

    Ok(Json(OAuthIdentity {
        connection_label: format!("PayRaider ({})", claims.username),
        user_id: claims.sub,
        username: claims.username,
        client_id: claims.client_id,
        scopes: claims.scopes,
    }))
}

/// Create OAuth routes
pub fn routes(db: SqlitePool) -> Router {
    Router::new()
        .route("/api/oauth/authorize", post(authorize))
        .route("/api/oauth/token", post(token))
        .route("/api/oauth/revoke", post(revoke))
        .route("/api/oauth/apps", get(list_apps))
        // #2127 — Zapier authentication.test
        .route("/api/oauth/me", get(me))
        .with_state(db)
}