# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PayRaider API",
    "description": "API for Stellar network analytics, anchor monitoring, and payment corridor insights",
    "contact": {
      "name": "PayRaider Team",
      "email": "support@payraider.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.payraider.io",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "POST /api/auth/refresh - Refresh access token",
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Token refreshed"
          },
          "401": {
            "description": "Invalid or expired token"
          }
        }
      }
    },
...
    "/api/oauth/apps": {
      "get": {
        "tags": [
          "OAuth"
        ],
        "summary": "GET /api/oauth/apps - List OAuth apps for authenticated user",
        "operationId": "list_apps",
        "responses": {
          "200": {
            "description": "List of OAuth apps",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOAuthAppsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/oauth/authorize": {
      "post": {
        "tags": [
          "OAuth"
        ],
        "summary": "POST /api/oauth/authorize - Request authorization code",
        "operationId": "authorize",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthAuthorizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authorization code generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAuthorizeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/oauth/revoke": {
      "post": {
        "tags": [
          "OAuth"
        ],
        "summary": "POST /api/oauth/revoke - Revoke an access token",
        "operationId": "revoke",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthRevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Token revoked"
          },
          "401": {
            "description": "Invalid client credentials"
          }
        }
      }
    },
    "/api/oauth/token": {
      "post": {
        "tags": [
          "OAuth"
        ],
        "summary": "POST /api/oauth/token - Exchange authorization code for tokens",
        "operationId": "token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Token issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Invalid client credentials"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ListOAuthAppsResponse": {
        "type": "object",
        "required": [
          "apps"
        ],
        "properties": {
          "apps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OAuthAppInfo"
            }
          }
        }
      },
...
      "OAuthAppInfo": {
        "type": "object",
        "description": "List OAuth Apps Response",
        "required": [
          "client_id",
          "app_name",
          "created_at"
        ],
        "properties": {
          "app_name": {
            "type": "string"
          },
          "client_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          }
        }
      },
      "OAuthAuthorizeRequest": {
        "type": "object",
        "description": "OAuth Authorization Request (for /api/oauth/authorize)",
        "required": [
          "client_id",
          "redirect_uri",
          "response_type",
          "scope",
          "state"
        ],
        "properties": {
          "client_id": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string"
          },
          "response_type": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        }
      },
      "OAuthAuthorizeResponse": {
        "type": "object",
        "description": "OAuth Authorization Response",
        "required": [
          "authorization_code",
          "state"
        ],
        "properties": {
          "authorization_code": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        }
      },
      "OAuthRevokeRequest": {
        "type": "object",
        "description": "OAuth Revoke Request",
        "required": [
          "access_token",
          "client_id",
          "client_secret"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "client_id": {
            "type": "string"
          },
          "client_secret": {
            "type": "string"
          }
        }
      },
      "OAuthTokenRequest": {
        "type": "object",
        "description": "OAuth Token Request (for /api/oauth/token)",
        "required": [
          "grant_type",
          "client_id",
          "client_secret"
        ],
        "properties": {
          "client_id": {
            "type": "string"
          },
          "client_secret": {
            "type": "string"
          },
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "grant_type": {
            "type": "string"
          },
          "redirect_uri": {
            "type": [
              "string",
              "null"
            ]
          },
          "refresh_token": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "RefreshTokenRequest": {
        "type": "object",
        "description": "Refresh token request",
        "required": [
          "refresh_token"
        ],
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
      "TokenResponse": {
        "type": "object",
        "description": "OAuth Token Response",
        "required": [
          "access_token",
          "refresh_token",
          "token_type",
          "expires_in",
          "scope"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer",
            "format": "int64"
          },
          "refresh_token": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "token_type": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}